PIN-2946 Resolved PR issues
