+++ src/main/scala/it/pagopa/interop/purposeprocess/api/impl/RiskAnalysisValidation.scala
@@ -11,7 +11,7 @@ import it.pagopa.interop.purposeprocess.error._
 import it.pagopa.interop.purposeprocess.model._
 import it.pagopa.interop.purposeprocess.model.riskAnalysisRules.{DependencyEntry, ValidationEntry}
 import it.pagopa.interop.purposeprocess.model.riskAnalysisTemplate._
-import it.pagopa.interop.purposeprocess.service.impl.RiskAnalysisServiceImpl
+import it.pagopa.interop.purposeprocess.service.RiskAnalysisService
 import it.pagopa.interop.tenantmanagement.client.model.TenantKind
 import spray.json._
 
@@ -27,7 +27,7 @@ object RiskAnalysisValidation {
   def validate(form: RiskAnalysisForm, schemaOnlyValidation: Boolean)(
     kind: TenantKind
   ): ValidationResult[RiskAnalysisFormSeed] = {
-    RiskAnalysisServiceImpl
+    RiskAnalysisService
       .riskAnalysisForms()
       .get(kind)
       .fold[ValidationResult[RiskAnalysisFormSeed]](MissingTenantKindConfiguration(kind).invalidNec)(
