+++ src/main/scala/it/pagopa/interop/purposeprocess/error/PurposeProcessErrors.scala
@@ -64,5 +64,8 @@ object PurposeProcessErrors {
       extends ComponentError("0019", s"Risk Analysis Configuration for Tenant ${tenantId.toString} not found")
 
   final case class RiskAnalysisConfigLatestVersionNotFound(tenantKind: TenantKind)
-      extends ComponentError("0020", s"Latest Risk Analysis Configuration for tenant kind ${tenantKind.toString} not found")
+      extends ComponentError(
+        "0020",
+        s"Latest Risk Analysis Configuration for tenant kind ${tenantKind.toString} not found"
+      )
 }
