+++ src/test/scala/it/pagopa/interop/purposeprocess/PDFCreatorSpec.scala
@@ -8,7 +8,7 @@ import it.pagopa.interop.purposemanagement.client.model.{
 import it.pagopa.interop.purposeprocess.error.RiskAnalysisTemplateErrors._
 import it.pagopa.interop.purposeprocess.model.riskAnalysisTemplate._
 import it.pagopa.interop.purposeprocess.service.impl.PDFCreatorImpl.setupData
-import it.pagopa.interop.purposeprocess.service.impl.RiskAnalysisServiceImpl
+import it.pagopa.interop.purposeprocess.service.RiskAnalysisService
 import org.scalatest.Assertion
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.wordspec.AnyWordSpecLike
@@ -21,7 +21,7 @@ class PDFCreatorSpec extends AnyWordSpecLike with SpecHelper {
   import PDFCreatorSpec._
 
   val testConfig: RiskAnalysisFormConfig =
-    RiskAnalysisServiceImpl.loadRiskAnalysisFormConfig("riskAnalysisTemplate/forms/test.json")
+    RiskAnalysisService.loadRiskAnalysisFormConfig("riskAnalysisTemplate/forms/test.json")
 
   val languages = List(LanguageIt, LanguageEn)
